requirements.txt
SQLAlchemy >= 1.4.36
psycopg2-binary >= 2.8.6
pandas >= 1.2.5
